New Homepage + Updated "well"
##New Homepage
I have created a much better and inviting homepage design. It still needs a lot of work but this at least is a little better than what we have. Since the "quickstart" guide is also given once you login and download the client it seemed redundant to have on the homepage.

The homepage uses inline styles for now maybe this will be updated in the future but I thought it would be better to just inline style instead of creating one time use styles in CSS while it's still in development.

Please provide additional fixes if you can, this is far from being "done"

## New Well Design
Well has been updated to a deep purple to better integrate with purple design throughout.